avoiding division by zero and usigned int underflow refs #1317
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@16508 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/netimport/NIImporter_VISUM.cpp
@@ -1000,12 +1000,12 @@ void NIImporter_VISUM::parse_LanesConnections() {
     if (fromLane - fromLaneOffset < 0) {
         fromLaneOffset = 0;
     } else {
-        fromLane = fromEdge->getNumLanes() - (fromLane - fromLaneOffset) - 1;
+        fromLane = (int)fromEdge->getNumLanes() - (fromLane - fromLaneOffset) - 1;
     }
     if (toLane - toLaneOffset < 0) {
         toLaneOffset = 0;
     } else {
-        toLane = toEdge->getNumLanes() - (toLane - toLaneOffset) - 1;
+        toLane = (int)toEdge->getNumLanes() - (toLane - toLaneOffset) - 1;
     }
     //
     if ((int) fromEdge->getNumLanes() <= fromLane) {
